// src/app/modules/result/models/student-bulk-entry.ts
export class StudentBulkEntry {
  
    STUDENT_ID: number;
    ORG_CODE: string;
    CAMPUS_CODE: string;
    CLASS_CODE: string;
    GROUP_CODE: string;
    SECTION_CODE: string;
    YEAR_CODE: string;
    SEMESTER_CODE: string;
    SHIFT_CODE: string;
    TERM_ID: number; 
    VERSION_CODE: string;
    START_SESSION_CODE: string;
    SESSION_CODE: string;
    CUR_SESSION_CODE: string;
    ADMISSION_DATE: any;
    CLASS_ROLL: any;
    STUDENT_NAME: any;
    FATHERS_NAME: any;
    SMS_MOBILE_NUM: any;
    RowStatus: number;
    USER_NAME: string;
    
}

export class StudentBulkEntryVM {
    STUDENT_ID: number;
    ORG_CODE: string;
    CAMPUS_CODE: string;
    CLASS_CODE: string;
    GROUP_CODE: string;
    SECTION_CODE: string;
    YEAR_CODE: string;
    SEMESTER_CODE: string;
    SHIFT_CODE: string;
    VERSION_CODE: string;
    START_SESSION_CODE: string;
    CUR_SESSION_CODE: string;
    ADMISSION_DATE: any;
    CLASS_ROLL: any;
    STUDENT_NAME: any;
    FATHERS_NAME: any;
    SMS_MOBILE_NUM: string;
    RowStatus: number;
    USER_NAME: string;
}