// src/app/modules/result/models/result-mark-chart-array.ts
export class ResultMarkChartArray{
    UNIQUE_ID: number;
    ORG_CODE: string;
    CAMPUS_CODE: string;
    TERM_ID: number; 
    CLASS_CODE: string; 
    SESSION_CODE: string; 
    GROUP_CODE: string; 
    VERSION_CODE: string; 
    SECTION_CODE: string; 
    YEAR_CODE: string; 
    SEMESTER_CODE: string; 
    TERM_DESCRIPTION: any[];
    SUBJECT_CODE: string[];
    SUBJECT_ID: any[];
    JOIN_SUBJECT_CODE: string[]; 
    SUBJECT_NAME: string[];
    CRITERIA: string[]; 
    CRITERIA_SHOW: string[]; 
    CRITERIA_PASS: number[]; 
    IS_JOIN: string[]; 
    CALCULATION_TYPE: string[]; 
    JOIN_ORDER: number[];
    SUBJECT_JOINING: any[]; 
    JOIN_PASS_MARK: number[]; 
    MULTIPLY_BY_JOIN: number[]; 
    TOTAL_SUBJECT: number[];
    EXAM_TOTAL_MARK: number[]; 
    MULTIPLY_BY: number[];
    TOTAL_PASS_MARK: number[]; 
    EXAM_1_TOTAL_MARK: number[]; 
    EXAM_1_PERCENT_MARK: number[]; 
    EXAM_1_PASS_MARK: number[]; 
    EXAM_2_TOTAL_MARK: number[]; 
    EXAM_2_PERCENT_MARK: number[]; 
    EXAM_2_PASS_MARK: number[];
    EXAM_3_TOTAL_MARK: number[]; 
    EXAM_3_PERCENT_MARK: number[]; 
    EXAM_3_PASS_MARK: number[]; 
    EXAM_4_TOTAL_MARK: number[];
    EXAM_4_PERCENT_MARK: number[]; 
    EXAM_4_PASS_MARK: number[];
    EXAM_5_TOTAL_MARK: number[]; 
    EXAM_5_PERCENT_MARK: number[]; 
    EXAM_5_PASS_MARK: number[];
    EXAM_6_TOTAL_MARK: number[]; 
    EXAM_6_PERCENT_MARK: number[];
    EXAM_6_PASS_MARK: number[]; 
    IS_EFFECTIVE: string;




    }